gcf/create-order/requirements.txt
functions-framework==3.*
google-cloud-storage==2.18.2
google-cloud-pubsub==2.23.1
grpcio==1.66.1
grpcio-tools==1.66.1
requests==2.32.3